CON-128: Make processed deploy cost type Long

+++ models/src/main/scala/io/casperlabs/models/InternalProcessedDeploy.scala
@@ -3,7 +3,6 @@ import io.casperlabs.casper.protocol.Deploy
 
 final case class InternalProcessedDeploy(
     deploy: Deploy,
-    //TODO: `cost` should be a Long, not a Double
-    cost: Double,
+    cost: Long,
     result: DeployResult
 )
